doc_processor/requirements.txt
# Core dependencies
firecrawl-py>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0

PyMuPDF==1.26.5
